Auto Add Magento v2.0.17

+++ dev/tests/integration/testsuite/Magento/Catalog/_files/product_group_prices_rollback.php
@@ -1,23 +1,21 @@
 <?php
 /**
- * Copyright © Magento, Inc. All rights reserved.
+ * Copyright © 2013-2017 Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
 
 /** @var \Magento\Framework\Registry $registry */
-$registry = \Magento\TestFramework\Helper\Bootstrap::getObjectManager()->get(\Magento\Framework\Registry::class);
+$registry = \Magento\TestFramework\Helper\Bootstrap::getObjectManager()->get('Magento\Framework\Registry');
 
 $registry->unregister('isSecureArea');
 $registry->register('isSecureArea', true);
 
-$repository = \Magento\TestFramework\Helper\Bootstrap::getObjectManager()->create(
-    \Magento\Catalog\Model\ProductRepository::class
-);
-try {
-    $product = $repository->get('simple', false, null, true);
+/** @var $product \Magento\Catalog\Model\Product */
+$product = \Magento\TestFramework\Helper\Bootstrap::getObjectManager()->create('Magento\Catalog\Model\Product');
+$product->load(1);
+if ($product->getId()) {
     $product->delete();
-} catch (\Magento\Framework\Exception\NoSuchEntityException $e) {
-    //Entity already deleted
 }
+
 $registry->unregister('isSecureArea');
 $registry->register('isSecureArea', false);
